fix: moved pages to pages directory
