+++ src/app/(pages)/payment/page.tsx
@@ -3,19 +3,19 @@ import Image from 'next/image'
 import React, { useEffect, useState } from 'react'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
-import CancelIcon from '../../../public/images/x-icon.svg'
-import AmexLogo from '../../../public/images/payments-logo/amex.svg'
-import ApplePayLogo from '../../../public/images/payments-logo/apple-pay.svg'
-import DiscoverLogo from '../../../public/images/payments-logo/discover.svg'
-import GooglePayLogo from '../../../public/images/payments-logo/google-pay.svg'
-import MasterCardLogo from '../../../public/images/payments-logo/mastercard.svg'
-import PaypalLogo from '../../../public/images/payments-logo/paypal.svg'
-import VisaLogo from '../../../public/images/payments-logo/visa.svg'
-import CardIcon from '../../../public/images/payments-logo/card.svg'
-import HelpIcon from '../../../public/images/payments-logo/help-icon.svg'
-import Check from '../../../public/images/check.svg'
+import CountryOptions from '../../components/country_option'
 
-import CountryOptions from '../components/country_option'
+import CancelIcon from '../../../../public/images/x-icon.svg'
+import AmexLogo from '../../../../public/images/payments-logo/amex.svg'
+import ApplePayLogo from '../../../../public/images/payments-logo/apple-pay.svg'
+import DiscoverLogo from '../../../../public/images/payments-logo/discover.svg'
+import GooglePayLogo from '../../../../public/images/payments-logo/google-pay.svg'
+import MasterCardLogo from '../../../../public/images/payments-logo/mastercard.svg'
+import PaypalLogo from '../../../../public/images/payments-logo/paypal.svg'
+import VisaLogo from '../../../../public/images/payments-logo/visa.svg'
+import CardIcon from '../../../../public/images/payments-logo/card.svg'
+import HelpIcon from '../../../../public/images/payments-logo/help-icon.svg'
+import Check from '../../../../public/images/check.svg'
 
 
 const Payment = () => {
