+++ src/app/(pages)/tournament_page/page.tsx
@@ -1,54 +1,54 @@
 "use client"
 import React, { useState } from 'react'
-import Links from '../components/links'
+import Links from '../../components/links'
 import { useRouter } from 'next/navigation'
-import { TournamentsListCard } from '../components/card'
+import { TournamentsListCard } from '../../components/card'
 import Image from 'next/image'
-
-import DollarIcon from '../../../public/images/dollar.svg'
-import CalendarIcon from '../../../public/images/calendar-icon.svg'
-import HashtagIcon from '../../../public/images/hash-02.svg'
-import PlayersIcon from '../../../public/images/users-01.svg'
-import HourGlassIcon from '../../../public/images/hour-glass.svg'
-import PrizeIcon from '../../../public/images/monetization-browser-bag.svg'
-import LiveGamesIcon from '../../../public/images/currency-dollar-circle.svg'
-import CameraPlus from '../../../public/images/camera-plus.svg'
-import GifIcon from '../../../public/images/image-document-gif.svg'
-import BellIcon from '../../../public/images/bell-02.svg'
-import ThumbsUp from '../../../public/images/thumbs-up.svg'
-import ThumbsDown from '../../../public/images/thumbs-down.svg'
-import CornerDownRight from '../../../public/images/corner-down-right.svg'
-import ThreeDots from '../../../public/images/dots-vertical.svg'
-import OpenWebLogo from '../../../public/images/open-web-logo.svg'
-import ArrowDown from '../../../public/images/arrow-down.svg'
-import DiagonalLines from '../../../public/images/green-diagonal.svg'
-import TransitionPattern from '../../../public/images/transition-pattern.svg'
-import BringATrailerLogo from '../../../public/images/bring-a-trailer-logo.svg'
-import ProfilePhoto from '../../../public/images/ellipse-415.svg'
-import CarFaxLogo from '../../../public/images/show-me-carfax.svg'
-import GiftIcon from '../../../public/images/gift-02.svg'
-import WatchListIcon from '../../../public/images/watchlist-icon.svg'
-import ThropyIconBlue from '../../../public/images/thropy-blue-big.svg'
-import CarsImage from '../../../public/images/cars-icon.svg'
-import CarIcon from '../../../public/images/car-01.svg'
-
-
-import PhotoOne from '../../../public/images/car-view-page/photoOne.svg'
-import PhotoTwo from '../../../public/images/car-view-page/photoTwo.svg'
-import PhotoThree from '../../../public/images/car-view-page/photoThree.svg'
-import PhotoFour from '../../../public/images/car-view-page/photoFour.svg'
-import PhotoFive from '../../../public/images/car-view-page/photoOne.svg'
-
-import AvatarOne from '../../../public/images/avatar-one.svg'
-import AvatarTwo from '../../../public/images/avatar-two.svg'
-import AvatarThree from '../../../public/images/avatar-three.svg'
-import AvatarFour from '../../../public/images/avatar-four.svg'
-import { LatestNews } from '../components/how_hammeshift_works'
-import { SubscribeSmall } from '../components/subscribe'
-import { TournamentsCard } from '../components/card'
-import Footer from '../components/footer'
+import { LatestNews } from '../../components/how_hammeshift_works'
+import { SubscribeSmall } from '../../components/subscribe'
+import { TournamentsCard } from '../../components/card'
+import Footer from '../../components/footer'
 import Link from 'next/link'
 import TournamentWagerPage from '../tournament_wager_page/page'
+
+import DollarIcon from '../../../../public/images/dollar.svg'
+import CalendarIcon from '../../../../public/images/calendar-icon.svg'
+import HashtagIcon from '../../../../public/images/hash-02.svg'
+import PlayersIcon from '../../../../public/images/users-01.svg'
+import HourGlassIcon from '../../../../public/images/hour-glass.svg'
+import PrizeIcon from '../../../../public/images/monetization-browser-bag.svg'
+import LiveGamesIcon from '../../../../public/images/currency-dollar-circle.svg'
+import CameraPlus from '../../../../public/images/camera-plus.svg'
+import GifIcon from '../../../../public/images/image-document-gif.svg'
+import BellIcon from '../../../../public/images/bell-02.svg'
+import ThumbsUp from '../../../../public/images/thumbs-up.svg'
+import ThumbsDown from '../../../../public/images/thumbs-down.svg'
+import CornerDownRight from '../../../../public/images/corner-down-right.svg'
+import ThreeDots from '../../../../public/images/dots-vertical.svg'
+import OpenWebLogo from '../../../../public/images/open-web-logo.svg'
+import ArrowDown from '../../../../public/images/arrow-down.svg'
+import DiagonalLines from '../../../../public/images/green-diagonal.svg'
+import TransitionPattern from '../../../../public/images/transition-pattern.svg'
+import BringATrailerLogo from '../../../../public/images/bring-a-trailer-logo.svg'
+import ProfilePhoto from '../../../../public/images/ellipse-415.svg'
+import CarFaxLogo from '../../../../public/images/show-me-carfax.svg'
+import GiftIcon from '../../../../public/images/gift-02.svg'
+import WatchListIcon from '../../../../public/images/watchlist-icon.svg'
+import ThropyIconBlue from '../../../../public/images/thropy-blue-big.svg'
+import CarsImage from '../../../../public/images/cars-icon.svg'
+import CarIcon from '../../../../public/images/car-01.svg'
+
+import PhotoOne from '../../../../public/images/car-view-page/photoOne.svg'
+import PhotoTwo from '../../../../public/images/car-view-page/photoTwo.svg'
+import PhotoThree from '../../../../public/images/car-view-page/photoThree.svg'
+import PhotoFour from '../../../../public/images/car-view-page/photoFour.svg'
+import PhotoFive from '../../../../public/images/car-view-page/photoOne.svg'
+
+import AvatarOne from '../../../../public/images/avatar-one.svg'
+import AvatarTwo from '../../../../public/images/avatar-two.svg'
+import AvatarThree from '../../../../public/images/avatar-three.svg'
+import AvatarFour from '../../../../public/images/avatar-four.svg'
+
 
 const CarViewData = {
     name: "13k-Mile 2011 Mercedes Benz SLS AMG",
