+++ src/app/(pages)/tournament_wager_page/page.tsx
@@ -1,16 +1,17 @@
 import React from 'react'
 import Link from 'next/link'
 import Image from 'next/image'
-import MoneyBag from '../../../public/images/money-bag-green.svg'
-import PlayersIcon from '../../../public/images/players-icon-green.svg'
-import HourGlass from '../../../public/images/hour-glass.svg'
-import Dollar from '../../../public/images/dollar.svg'
-import CancelIcon from '../../../public/images/x-icon.svg'
-import SedanPhotoOne from '../../../public/images/tournament-wager/sedan-photo-one.svg'
-import SedanPhotoTwo from '../../../public/images/tournament-wager/sedan-photo-two.svg'
-import SedanPhotoThree from '../../../public/images/tournament-wager/sedan-photo-three.svg'
-import SedanPhotoFour from '../../../public/images/tournament-wager/sedan-photo-four.svg'
-import SedanPhotoFive from '../../../public/images/tournament-wager/sedan-photo-five.svg'
+
+import MoneyBag from '../../../../public/images/money-bag-green.svg'
+import PlayersIcon from '../../../../public/images/players-icon-green.svg'
+import HourGlass from '../../../../public/images/hour-glass.svg'
+import Dollar from '../../../../public/images/dollar.svg'
+import CancelIcon from '../../../../public/images/x-icon.svg'
+import SedanPhotoOne from '../../../../public/images/tournament-wager/sedan-photo-one.svg'
+import SedanPhotoTwo from '../../../../public/images/tournament-wager/sedan-photo-two.svg'
+import SedanPhotoThree from '../../../../public/images/tournament-wager/sedan-photo-three.svg'
+import SedanPhotoFour from '../../../../public/images/tournament-wager/sedan-photo-four.svg'
+import SedanPhotoFive from '../../../../public/images/tournament-wager/sedan-photo-five.svg'
 
 const TournamentWagerPage = () => {
     const tournamentWagerPageData = [{
